Refactor type checker
Because constants can depend on declared types and vice versa and because
variables can depend on function declarations and vice versa, the new package
handler orchestrates the type checking operations for all of these. This means
first finding all actions that need to be taken, then ordering them to respect
dependecies (and reporting any loops), and finally executing the ordered
actions.

The type handler deals with all type expressions and constructs their
corresponding types::Type. It also handles type declarations.

The constant handler deals with all constant expressions and declarations.

The variable handler deals with variable definitions, including creating the
init order for the package.

The expression handler is going to type check expressions and along with other
handlers will be used in the future to type check statements and function
bodies.