Refactor types package
types::Info now stores ExprInfo objects for each AST expression instead of the
previous (somewhat) parallel maps for expression kinds and types. In the
future, the constant value of an expression should also be added to ExprInfo,
if known. This requires refactoring of the constant_handler in the type
checker though.
A couple cc files, containing only getter method were removed. The Selection
and Initializer are now handled with value semantics.