Generate underlying types for type instances lazily
This finally fixes the issue of missing underlying types for type instances.
Underlying types of type instances can not be computed ahead of time because it
is possible for the expansions to be unterminated, e.g.:

type List<T> interface {
    () AllPossibleOrderings() List<List<T>>
}

Here, the result of AllPossibleOrderings again has a method returning
List<List<List<T>>> and so on. Therefore underlying types have to be generated
as needed, when the utility types::Underlying function gets called. This
function now requires a types::InfoBuilder argument to make sure newly
generated types are owned by a types::Info instance.
types::NamedType now caches underlying types for all type arguments for which
they have been generated.
This helped to resolve several warnings. test2 has no more warnings now and a
bug in the test itself was found and fixed.