Move x86_64 test from tests.cc to dedicated unit test setup
This defines a new execution_test_executable target that generates x86_64
machine code, writes it to a memory page with write and execute permissions,
and invokes it. The execution_test target (written in Go) takes this binary,
runs it as a subcommand, and tests that the output written to stdout matches
the expected output.
Previously, the inspection of the output was done manually.