Implement memory instrs and fix bugs in IR interpreter
Implement pointer offset, niltest, malloc, load, store, and free in the
interpreter. Fix bugs in the main run loop that caused incorrect continuation
after jump instructions. Fix bugs that caused updates to computed values to
not be written to FuncContext.
Add new testcase that sums numbers from 0 to 9 (result is 45).